terraform/function/requirements.txt
# Cloud Function requirements for Titanic data loader
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-storage==2.*
pandas==2.*
